Allow python-snappy to be compiled in Windows
Since MSVC only allows the inline keyword to be used on .cpp files, it is needed to check if _MSC_VER is defined, so it can be detected if the compiler is MSVC and use the vendor-specific __inline keyword instead. Without this change, the only way to compile python-snappy is to download the source code, make the change and compile/install manually.