StreamDecompressor: faster decompression via a lot less memcpy
The decompression loop does a lot of copying of memory. Performance
tests indicate more time is spend in copying memory than in the actual
snappy decompression algorithm.

Use memoryviews and bytearrays to reduce the number of memcpys. This
gives about 24% faster execution on my test setup.

NOTE: the interface is changed a bit: the decompressor returns a
memoryview() into a bytearray() instead of a bytes() object. This may
break existing code.